+++ phpmyfaq/assets/src/vendors.js
@@ -17,7 +17,6 @@ import $ from 'jquery';
 window.jQuery = $;
 window.$ = $;
 
-require('popper.js');
 require('bootstrap');
 require('bootstrap-3-typeahead');
 require('bootstrap-datepicker');